requirements.txt
simplejson
lxml
codecov
coverage
flask
pre - commit
pytest
PyJWT
requests
sqlalchemy